Remove dependency on ppx_deriving
The ppx_deriving package doesn't actually contain a library named
"ppx_deriving", but it does contain several libraries with
"ppx_deriving" as a prefix (e.g. ppx_deriving.eq, ppx_deriving.ord).
Currently it's possible to depend on library prefixes like
"ppx_deriving" when the package containing the library is installed
globally, but not if the package is vendored. That is, if the package
ppx_deriving were vendored, dune would be unable to find a library named
"ppx_deriving".